backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.19
pydantic==2.1.1
alembic==1.11.1
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pandas==2.0.3
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1